moq-mux: drop deprecated stubs and inline msf into catalog
- Delete hang::container::OrderedConsumer / OrderedFrame deprecation
  stubs. They only panicked; all in-tree callers were migrated to
  moq_mux::export::OrderedConsumer in the previous backport.
- Inline rs/moq-mux/src/msf.rs (to_msf + tests) into import/catalog.rs.
  CatalogGuard::drop is the only caller; to_msf was pub but unused
  outside the crate. Removes the bare top-level msf module.
- Drop deprecated type aliases in import/{framed,stream}.rs:
  DecoderFormat, Decoder, StreamDecoder. None used in-tree.
