+++ rs/hang/src/container/mod.rs
@@ -1,8 +1,5 @@
-mod consumer;
 mod frame;
 mod producer;
 
-#[allow(deprecated)]
-pub use consumer::*;
 pub use frame::*;
 pub use producer::*;
